unzip epub

+++ controller/open_service.py
@@ -345,16 +345,19 @@ class OpenService(BaseService):
             if zf:
                 zf.close()
 
-    def find_file(self, file_start, root_dir):
+    def find_file(self, file_starts, root_dir):
         import os
         cover_file_path = None
         if os.path.exists(root_dir):
             find = False
             for root, sub_dirs, files in os.walk(root_dir):
                 for special_file in files:
-                    if special_file.startswith(file_start):
-                        cover_file_path = os.path.join(root_dir, special_file)
-                        find = True
+                    for _start in file_starts:
+                        if special_file.startswith(_start):
+                            cover_file_path = os.path.join(root_dir, special_file)
+                            find = True
+                            break
+                    if find:
                         break
                 if find:
                     break
@@ -400,37 +403,41 @@ class OpenService(BaseService):
                         current_dest_dir = os.path.join(dest_dir, sb.code)
                         if not os.path.exists(current_dest_dir):
                             os.makedirs(current_dest_dir)
-                        try:
-                            self.unzip_single(file_path, current_dest_dir)
-                            cover_dir = os.path.join(current_dest_dir, "OPS/images/")
-                            cover_file_path = self.find_file("cover.", cover_dir)
-                            if not cover_file_path:
-                                cover_dir = os.path.join(current_dest_dir, "OPS/")
-                                cover_file_path = self.find_file("cover.", cover_dir)
-                                if not cover_file_path:
-                                    cover_file_path = self.find_file("cover.", current_dest_dir)
-                            opf_dir = os.path.join(current_dest_dir, "OPS/")
-                            opf_file_path = self.find_file_by_end(".opf", opf_dir)
-                            if not opf_file_path:
-                                opf_file_path = self.find_file_by_end(".opf", current_dest_dir)
-                            params = {"pin": 1, "unziped": 1}
-                            if cover_file_path:
-                                params["cover"] = cover_file_path
-                            if opf_file_path:
-                                params["opf"] = opf_file_path
-                            print("unzip ok, name:", sb.name)
-                            StudyDao.update_books_by_id(params, sb.id)
-                            # del file
-                            os.remove(file_path)
-                        except Exception:
-                            need_up_unziped.append(sb.code)
-                            os.remove(file_path)
+                        ops_dir = os.path.join(current_dest_dir, "OPS/")
+                        if not os.path.exists(ops_dir):
+                            ops_dir = os.path.join(current_dest_dir, "OEBPS/")
+                        if os.path.exists(ops_dir):
                             try:
-                                if os.path.exists(current_dest_dir):
-                                    os.rmdir(current_dest_dir)
+                                self.unzip_single(file_path, current_dest_dir)
+                                cover_dir = os.path.join(ops_dir, "images/")
+                                cover_file_path = self.find_file(["cover.j", "cover.png"], cover_dir)
+                                if not cover_file_path:
+                                    cover_file_path = self.find_file(["cover.j", "cover.png"], ops_dir)
+                                    if not cover_file_path:
+                                        cover_file_path = self.find_file(["cover.j", "cover.png"], current_dest_dir)
+                                opf_file_path = self.find_file_by_end(".opf", ops_dir)
+                                if not opf_file_path:
+                                    opf_file_path = self.find_file_by_end(".opf", current_dest_dir)
+                                params = {"pin": 1, "unziped": 1}
+                                if cover_file_path:
+                                    params["cover"] = cover_file_path
+                                if opf_file_path:
+                                    params["opf"] = opf_file_path
+                                print("unzip ok, name:", sb.name)
+                                StudyDao.update_books_by_id(params, sb.id)
+                                # del file
+                                os.remove(file_path)
                             except Exception:
-                                logger.error("remove err epud extract dir [{}] failed!".format(current_dest_dir),
-                                             exec_info=True)
+                                need_up_unziped.append(sb.code)
+                                os.remove(file_path)
+                                try:
+                                    if os.path.exists(current_dest_dir):
+                                        os.rmdir(current_dest_dir)
+                                except Exception:
+                                    logger.error("remove err epud extract dir [{}] failed!".format(current_dest_dir),
+                                                 exec_info=True)
+                        else:
+                            StudyDao.update_books_by_id({"pin": 3}, sb.id)
 
                 if need_up_unziped:
                     print("will batch_update_books_by_codes:", need_up_unziped)
